打分及util重构&elutionScore攻坚

+++ src/main/java/com/westlake/air/pecs/domain/bean/score/IntegrateWindowMzIntensity.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class IntegrateWindowMzIntensity {
 
-    double mz;
+    float mz;
 
-    double intensity;
+    float intensity;
 
     boolean signalFound;
 }
